# docs/openapi.yaml
...
openapi: 3.1.1
info:
    title: Bluebell API
    version: 1.0.0
    description: |
        Bluebell 前后端联调接口，以当前后端实现为准，共 17 个接口。

        ## 鉴权

        发帖、删帖、查询和设置投票状态、发表评论和删除评论需要 Bearer JWT（`access_token`）。
        注册、登录、刷新令牌、退出登录、社区与帖子查询（含评论列表）、`/ping` 和 `/health` 无需 Bearer 令牌。
        公开接口不解析 `Authorization`，携带过期或无效令牌也不会返回 401。

        登录成功后，响应体返回短期 `access_token`（默认 15 分钟，由 `auth.access_token_ttl` 配置），
        同时通过 HttpOnly Cookie `refresh_token` 下发刷新令牌（默认 7 天，由 `auth.refresh_token_ttl` 配置）；
        刷新令牌不出现在响应体中。`access_token` 过期后调用 `POST /api/v1/refresh` 获取新的 `access_token`，
        刷新令牌每次使用即轮换，旧令牌立即失效；轮换不会延长会话，新令牌继承原令牌的剩余有效期，
        到期后需要重新登录。

        ## 响应格式

        请求体使用 JSON。业务 JSON 响应统一为 `{code, message, data}`：
        成功时 `code` 为 `OK`，失败时 `data` 为 `null`；前端按 HTTP 状态码和 `code` 判断结果。
        删除成功和退出登录返回 HTTP 204，无响应体；两个健康检查接口直接返回各自的 JSON 对象。

        `created_at` 为带时区的 RFC 3339 时间字符串，可能包含小数秒。
        社区、帖子、评论 ID 为正整数；`author_id` 为用户 UUID 字符串。

        ## 帖子与投票

        帖子列表不返回正文、赞成／反对票数或当前用户的投票状态；详情返回正文和票数，当前用户的投票状态通过独立接口查询。
        投票成功后，详情票数和当前用户的投票状态可直接查询；帖子列表的可见性、总数和热度排序异步更新，可能短暂滞后。

        ## 跨域

        后端未配置 CORS。浏览器中的前端或文档页面跨域调用时，需要通过开发代理或部署网关处理，导入本文件不会开启跨域支持。
        依赖 Cookie 的 `refresh` 和 `logout` 接口还要求前端与后端同站，或由代理转发使 Cookie 作用于前端域名。
servers:
    - url: http://localhost:8080
      description: 本地开发默认地址；联调时替换为实际服务地址，路径已包含 /api/v1
...
security:
    - bearerAuth: []
paths:
...
    /api/v1/refresh:
        post:
            tags: [认证]
            operationId: refreshToken
            summary: 刷新访问令牌
            description: |
                使用登录时下发的 refresh_token Cookie 换取新的 access_token，无请求体，也不需要 Authorization 请求头。
                刷新令牌一次有效：成功后服务端立即作废旧令牌，并通过 Set-Cookie 下发新令牌。
                新令牌继承旧令牌的剩余有效期，不会延长会话；到期后需要重新登录。
                未携带 Cookie、令牌不存在、已过期或已被使用（包括重放旧令牌）均返回 UNAUTHORIZED。
                前端收到 401 后应清除本地状态并跳转登录。
            security:
                - refreshCookie: []
            responses:
                "200":
                    description: 刷新成功，响应体只包含新的 access_token，不包含 username。
                    headers:
                        Set-Cookie:
                            $ref: "#/components/headers/SetRefreshTokenCookie"
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: "#/components/schemas/SuccessResponse"
                                    - type: object
                                      properties:
                                          data:
                                              type: object
                                              required:
                                                  [access_token, token_type]
                                              properties:
                                                  access_token:
                                                      type: string
                                                      description: 新的 JWT 访问令牌；示例为占位值。
                                                  token_type:
                                                      type: string
                                                      const: Bearer
                            example:
                                code: OK
                                message: 成功
                                data:
                                    access_token: "<access_token>"
                                    token_type: Bearer
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "500":
                    $ref: "#/components/responses/InternalError"
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 登录后获取 access_token，请求头为 Authorization：Bearer <access_token>。
        refreshCookie:
            type: apiKey
            in: cookie
            name: refresh_token
            description: 登录时由服务端通过 HttpOnly Cookie 下发，浏览器自动携带；非浏览器客户端需要自行保存并回传。
    headers:
        SetRefreshTokenCookie:
            description: |
                下发或轮换 refresh_token。属性：HttpOnly、Path=/、SameSite=Lax；
                Max-Age 为刷新令牌剩余有效期；Secure 由 auth.cookie_secure 决定，生产环境需开启 HTTPS 并设为 true。
            schema:
                type: string
            example: refresh_token=<refresh_token>; Path=/; Max-Age=604800; HttpOnly; SameSite=Lax
...
    schemas:
...
        SuccessResponse:
            type: object
            required: [code, message, data]
            properties:
                code:
                    type: string
                    const: OK
                message:
                    type: string
                    const: 成功
                data:
                    description: 业务数据，具体结构见各接口。
        ErrorResponse:
            type: object
            required: [code, message, data]
            description: |
                业务失败响应，data 固定为 null。各接口只返回其 responses 中列出的业务错误。
                同一 HTTP 状态码可能对应不同 code，具体含义见各响应的描述与示例。
            properties:
                code:
                    type: string
                    enum:
                        - INVALID_PARAMS
                        - USERNAME_EXISTS
                        - INVALID_CREDENTIALS
                        - UNAUTHORIZED
                        - FORBIDDEN
                        - COMMUNITY_NOT_FOUND
                        - POST_NOT_FOUND
                        - COMMENT_NOT_FOUND
                        - INTERNAL_ERROR
                message:
                    type: string
                    description: 业务错误对应的中文提示，不含具体字段校验详情。
                data:
                    type: "null"
...
    responses:
...
        Unauthorized:
            description: |
                UNAUTHORIZED — 未登录或登录已过期。
                Bearer 鉴权接口：Authorization 缺失、格式错误或 access_token 无效或已过期，鉴权先于业务参数校验。
                刷新接口：refresh_token Cookie 缺失、无效、已过期或已被轮换。
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    example:
                        code: UNAUTHORIZED
                        message: 未登录或登录已过期
                        data: null
...
        InternalError:
            description: INTERNAL_ERROR — 业务处理发生服务器内部错误。
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    example:
                        code: INTERNAL_ERROR
                        message: 服务器内部错误
                        data: null